Move the CollectVCSFixCommitPipeline base pipelines to pipes Remove CollectOpensslFixCommits Pipeline
